+++ source/compiler-core/slang-visual-studio-compiler-util.cpp
@@ -12,7 +12,7 @@
 #endif
 
 #include "../core/slang-io.h"
-#include "slang-artifact-info.h"
+#include "slang-artifact-desc-util.h"
 
 namespace Slang
 {
@@ -264,15 +264,15 @@ namespace Slang
     {
         auto desc = artifact->getDesc();
 
-        if (ArtifactInfoUtil::isCpuBinary(desc) && desc.kind == ArtifactKind::Library)
+        if (ArtifactDescUtil::isCpuBinary(desc) && desc.kind == ArtifactKind::Library)
         {
             // Get the libray name and path
             ComPtr<IFileArtifactRepresentation> fileRep;
             SLANG_RETURN_ON_FAIL(artifact->requireFile(ArtifactKeep::No, fileRep.writeRef()));
 
-            libPathPool.add(ArtifactInfoUtil::getParentPath(fileRep));
+            libPathPool.add(ArtifactDescUtil::getParentPath(fileRep));
             // We need the extension for windows
-            cmdLine.addArg(ArtifactInfoUtil::getBaseName(artifact->getDesc(), fileRep) + ".lib");
+            cmdLine.addArg(ArtifactDescUtil::getBaseName(artifact->getDesc(), fileRep) + ".lib");
         }
     }
