+++ source/slang/slang-compiler.cpp
@@ -13,7 +13,8 @@
 #include "slang-compiler.h"
 
 #include "../compiler-core/slang-lexer.h"
-#include "../compiler-core/slang-artifact.h"
+
+#include "../compiler-core/slang-artifact-desc-util.h"
 #include "../compiler-core/slang-artifact-representation.h"
 
 #include "slang-lower-to-ir.h"
@@ -57,7 +58,7 @@ namespace Slang
 
 bool isHeterogeneousTarget(CodeGenTarget target)
 {
-    return ArtifactDesc::makeFromCompileTarget(asExternal(target)).style == ArtifactStyle::Host;
+    return ArtifactDescUtil::makeDescFromCompileTarget(asExternal(target)).style == ArtifactStyle::Host;
 }
 
 void printDiagnosticArg(StringBuilder& sb, CodeGenTarget val)
@@ -997,7 +998,7 @@ void printDiagnosticArg(StringBuilder& sb, CodeGenTarget val)
 
     static bool _isCPUHostTarget(CodeGenTarget target)
     {
-        auto desc = ArtifactDesc::makeFromCompileTarget(asExternal(target));
+        auto desc = ArtifactDescUtil::makeDescFromCompileTarget(asExternal(target));
         return desc.style == ArtifactStyle::Host;
     }
 
@@ -1303,7 +1304,7 @@ void printDiagnosticArg(StringBuilder& sb, CodeGenTarget val)
 
         // If we aren't using LLVM 'host callable', we want downstream compile to produce a shared library
         if (compilerType != PassThroughMode::LLVM && 
-            ArtifactDesc::makeFromCompileTarget(asExternal(target)).kind == ArtifactKind::HostCallable)
+            ArtifactDescUtil::makeDescFromCompileTarget(asExternal(target)).kind == ArtifactKind::HostCallable)
         {
             target = CodeGenTarget::ShaderSharedLibrary;
         }
@@ -1804,7 +1805,7 @@ void printDiagnosticArg(StringBuilder& sb, CodeGenTarget val)
                 ComPtr<ISlangBlob> blob;
                 if (SLANG_FAILED(result.getBlob(blob)))
                 {
-                    if (ArtifactDesc::makeFromCompileTarget(asExternal(targetReq->getTarget())).kind == ArtifactKind::HostCallable)
+                    if (ArtifactDescUtil::makeDescFromCompileTarget(asExternal(targetReq->getTarget())).kind == ArtifactKind::HostCallable)
                     {
                         // Some HostCallable are not directly representable as a 'binary'.
                         // So here, we just ignore if that appears the case, and don't output an unexpected error.
@@ -2448,7 +2449,7 @@ void printDiagnosticArg(StringBuilder& sb, CodeGenTarget val)
             return;
 
         auto target = getTargetFormat();
-        const auto desc = ArtifactDesc::makeFromCompileTarget(asExternal(target));
+        const auto desc = ArtifactDescUtil::makeDescFromCompileTarget(asExternal(target));
 
         if (desc.kind == ArtifactKind::Text)
         {
